add connect for app
