+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import {createStore, applyMiddleware} from 'redux';
 import {composeWithDevTools} from 'redux-devtools-extension';
 import {Provider} from "react-redux";
-import App from "./components/app/app.jsx";
+import App from "./components/app/app.connect.js";
 import reducer from './reducer/reducer.js';
 import thunk from 'redux-thunk';
 import createAPI from './api.js';
